Argus: rearrange layout, adding TopWidget and simplifying
