+++ uscope/gui/imager.py
