+++ uscope/gui/input_widget.py
